Trait whatwg_infra::surrogates::InfraUtf16Surrogate

pub trait InfraUtf16Surrogate {
    const LEADING_SURROGATE_MIN: u16;
    const LEADING_SURROGATE_MAX: u16;
    const TRAILING_SURROGATE_MIN: u16;
    const TRAILING_SURROGATE_MAX: u16;
    const SURROGATE_MIN: u16;
    const SURROGATE_MAX: u16;

    // Required methods
    fn is_surrogate_utf16(self) -> bool;
    fn is_leading_surrogate_utf16(self) -> bool;
    fn is_trailing_surrogate_utf16(self) -> bool;
}

Methods from the WHATWG Infra Standard for UTF-16 surrogates

Required Associated Constants

const LEADING_SURROGATE_MIN: u16

The minimum UTF-16 codepoint that can be represented as a leading surrogate

const LEADING_SURROGATE_MAX: u16

The minimum UTF-16 codepoint that can be represented as a leading surrogate

const TRAILING_SURROGATE_MIN: u16

The minimum UTF-16 codepoint that can be represented as a trailing surrogate

const TRAILING_SURROGATE_MAX: u16

The maximum UTF-16 codepoint that can be represented as a trailing surrogate

const SURROGATE_MIN: u16

The minimum UTF-16 codepoint that can be represented as a surrogate

const SURROGATE_MAX: u16

The maximum UTF-16 codepoint that can be represented as a surrogate

Required Methods

fn is_surrogate_utf16(self) -> bool

See the documentation for is_surrogate_utf16()

fn is_leading_surrogate_utf16(self) -> bool

See the documentation for is_leading_surrogate_utf16()

fn is_trailing_surrogate_utf16(self) -> bool

See the documentation for is_trailing_surrogate_utf16()

Object Safety

This trait is not object safe.